fix: separate repo mirror zips per branch
When pipeline was changed and re-started,
it took latest repository zip version.
This could be a commit not from the default branch.
In effect, the main pipeline could deploy a branch version.

Now, every branch is mirrored to a separate zip.
It prevents such collisions and also makes easier to manually
start main pipeline or branch builds by providing file name,
not exact object version.